refactor: rename directory for createPlayWrightUserAccess and remove unused file
